Initial commit for the csvCombiner repository.
Running the main method in CSVCombiner allows the first argument (a CSV file) to be output having combined all the remaining arguments (all also CSV files) into one file. This combining is done on the understanding that every CSV file makes use of the same ID scheme (and hence, entity identification can be done across files).
Entry acts as a wrapper for a HashMap, though future capabilities are possible.